move in memory states to state package
