+++ be1-go/internal/handler/messagedata/chirp/chirp_test.go
@@ -10,7 +10,7 @@ import (
 	mock2 "popstellar/internal/mock"
 	"popstellar/internal/mock/generator"
 	"popstellar/internal/repository"
-	"popstellar/internal/types"
+	"popstellar/internal/state"
 	"popstellar/internal/validation"
 	"strings"
 	"testing"
@@ -28,7 +28,7 @@ func Test_handleChannelChirp(t *testing.T) {
 		contains    string
 	}
 
-	subs := types.NewSubscribers()
+	subs := state.NewSubscribers()
 
 	db := mock2.NewRepository(t)
 
@@ -45,7 +45,7 @@ func Test_handleChannelChirp(t *testing.T) {
 	serverSecretKey := crypto.Suite.Scalar().Pick(crypto.Suite.RandomStream())
 	serverPublicKey := crypto.Suite.Point().Mul(serverSecretKey, nil)
 
-	conf := types.CreateConfig(ownerPublicKey, serverPublicKey, serverSecretKey, "clientAddress", "serverAddress")
+	conf := state.CreateConfig(ownerPublicKey, serverPublicKey, serverSecretKey, "clientAddress", "serverAddress")
 
 	chirp := New(conf, subs, db, schema)
