+++ be1-go/internal/handler/messagedata/election/election_test.go
@@ -11,6 +11,7 @@ import (
 	"popstellar/internal/message/query/method/message"
 	mock2 "popstellar/internal/mock"
 	"popstellar/internal/mock/generator"
+	"popstellar/internal/state"
 	"popstellar/internal/types"
 	"popstellar/internal/validation"
 	"testing"
@@ -32,7 +33,7 @@ func Test_handleChannelElection(t *testing.T) {
 
 	var args []input
 
-	subs := types.NewSubscribers()
+	subs := state.NewSubscribers()
 
 	db := mock2.NewRepository(t)
 
@@ -49,7 +50,7 @@ func Test_handleChannelElection(t *testing.T) {
 	serverSecretKey := crypto.Suite.Scalar().Pick(crypto.Suite.RandomStream())
 	serverPublicKey := crypto.Suite.Point().Mul(serverSecretKey, nil)
 
-	conf := types.CreateConfig(ownerPublicKey, serverPublicKey, serverSecretKey, "clientAddress", "serverAddress")
+	conf := state.CreateConfig(ownerPublicKey, serverPublicKey, serverSecretKey, "clientAddress", "serverAddress")
 
 	election := New(conf, subs, db, schema)
