+++ be1-go/internal/state/config.go
@@ -1,4 +1,4 @@
-package types
+package state
 
 import (
 	"encoding/base64"
