+++ be1-go/internal/state/hub_params.go
@@ -1,4 +1,4 @@
-package types
+package state
 
 import (
 	"popstellar/internal/network/socket"
