+++ be1-go/internal/state/queries.go
@@ -1,4 +1,4 @@
-package types
+package state
 
 import (
 	"popstellar/internal/errors"
