+++ be1-go/internal/state/sockets.go
@@ -1,4 +1,4 @@
-package types
+package state
 
 import (
 	"fmt"
