+++ be1-go/internal/state/subscribers.go
@@ -1,4 +1,4 @@
-package types
+package state
 
 import (
 	"encoding/json"
